+++ Software/StoryWatch/StoryWatch/UserControls/UCMediaListBox.xaml.cs
@@ -417,6 +417,13 @@ namespace StoryWatch.UserControls
 
         private void btnDeleteList_Click(object sender, RoutedEventArgs e)
         {
+            MessageBoxResult result = MessageBox.Show("Are you sure you want to delete this list?", "Delete", MessageBoxButton.YesNo, MessageBoxImage.Question);
+
+            if (result == MessageBoxResult.No)
+            {
+                return;
+            }
+
             listCategoryServices.DeleteListCategory(listCategory, StateManager.LoggedUser, StateManager.CurrentMediaCategory);
             GuiManager.OpenContent(new UCMediaHome(StateManager.CurrentMediaCategory));
         }
